chore: Move AppEnvironment into separate file
